Save Script Now Functions
The Script designed to save the game has now been implemented and confirmed working. Saves the "Last Meaningful Login" (to be determined as when the player actually does something to care for the friend) as a string, and when loading, parses the string into a proper DateTime object for the code to read. Documentation on this script exists inside the script to allow for easy reading and public access outside of the script.