Cleanup Minigame Base Created
Created the main core of the Clean-up Minigame. Placeholder graphics are used, and it needs some polishing before I'm content with it in this state.

Also, added a generic UI for the three gauges that needs a bit more tinkering to get it to work fully. Did some preemptive cleanup of the Unity Scene and made some scripts in preparation of actually writing the other two minigames.